Split the port helpers out and keep stop failures visible
Ports and start expectations move to their own file, bringing the
shared helpers back under the size limit. Stopping a mock now happens
outside the catch that expects the start to fail, so a stop that goes
wrong is not read as the failure the test wanted.
