+++ test/test-utils/stripe-mock/ports.test.ts
@@ -4,7 +4,7 @@ import {
   PORT_STEAL_TRIES,
   retryWhilePortTaken,
   startFailedOrPortTaken,
-} from "#test/test-utils/stripe-mock/helpers.ts";
+} from "#test/test-utils/stripe-mock/ports.ts";
 
 /** Stands in for a mock that started, which is what a taken port looks like. */
 const startedMock = () =>
